+++ clickhouse_connect/datatypes/container.py
@@ -1,11 +1,10 @@
 import array
 from typing import Dict, Sequence, MutableSequence, Any
 
-from clickhouse_connect.datatypes.base import UnsupportedType, ClickHouseType, TypeDef, EMPTY_TYPE_DEF
-from clickhouse_connect.datatypes.string import String
-from clickhouse_connect.driver.common import read_leb128, to_leb128, array_column, must_swap, write_uint64
+from clickhouse_connect import json_impl
+from clickhouse_connect.datatypes.base import ClickHouseType, TypeDef
+from clickhouse_connect.driver.common import read_leb128, to_leb128, array_column, must_swap
 from clickhouse_connect.datatypes.registry import get_from_name
-from clickhouse_connect.json_impl import json_impl
 
 
 class Array(ClickHouseType):
@@ -116,17 +115,17 @@ class Tuple(ClickHouseType):
             loc = e_type.read_native_prefix(source, loc)
         return loc
 
-    def read_native_data(self, source: Sequence, loc: int, num_rows: int, use_none = True):
+    def read_native_data(self, source: Sequence, loc: int, num_rows: int, use_none=True):
         columns = []
         e_names = self.element_names
         for e_type in self.element_types:
             column, loc = e_type.read_native_data(source, loc, num_rows, use_none)
             columns.append(column)
-        if e_names and self.read_format != 'tuple':
+        if e_names and self.read_format() != 'tuple':
             dicts = [{} for _ in range(num_rows)]
             for ix, x in enumerate(dicts):
-                for n, key in enumerate(e_names):
-                    x[key] = columns[n][ix]
+                for y, key in enumerate(e_names):
+                    x[key] = columns[y][ix]
             return dicts, loc
         return tuple(zip(*columns)), loc
 
@@ -176,7 +175,7 @@ class Map(ClickHouseType):
         return loc
 
     # pylint: disable=too-many-locals
-    def read_native_data(self, source: Sequence, loc: int, num_rows: int, use_none = True):
+    def read_native_data(self, source: Sequence, loc: int, num_rows: int, use_none=True):
         offsets, loc = array_column('Q', source, loc, num_rows)
         total_rows = offsets[-1]
         keys, loc = self.key_type.read_native_data(source, loc, total_rows, use_none)
@@ -252,19 +251,20 @@ class JSON(ClickHouseType):
     python_type = dict
 
     def _to_row_binary(self, value: Any, dest: MutableSequence):
-        value = bytes(json_impl.dumps(value))
+        value = json_impl.any_to_json(value)
         dest += to_leb128(len(value)) + value
 
     def _from_row_binary(self, source: Sequence, loc: int):
-        length, loc = read_leb128(source, loc)
-        return json_impl.loads(str(source[loc:loc + length])), loc + length
+        # ClickHouse will never return JSON/Object types, just tuples
+        return None, 0
 
     def write_native_prefix(self, dest: MutableSequence):
         dest.append(0x01)
 
+    # pylint: disable=duplicate-code
     def write_native_data(self, column: Sequence, dest: MutableSequence):
         app = dest.append
-        to_json = json_impl.dumps
+        to_json = json_impl.any_to_json
         for x in column:
             v = to_json(x)
             sz = len(v)