+++ clickhouse_connect/datatypes/format.py
@@ -69,13 +69,8 @@ def _convert_arguments(*args, **kwargs) -> Dict[str, str]:
 
 
 def _matching_types(pattern: str, fmt: str = None) -> Sequence[Type[ClickHouseType]]:
-    if '*' in pattern:
-        re_pattern = re.compile(pattern.replace('*', '.*'), re.IGNORECASE)
-        matches = [ch_type for type_name, ch_type in type_map.items() if re_pattern.match(type_name)]
-    elif pattern in type_map:
-        matches = [type_map[pattern]]
-    else:
-        matches = []
+    re_pattern = re.compile(pattern.replace('*', '.*'), re.IGNORECASE)
+    matches = [ch_type for type_name, ch_type in type_map.items() if re_pattern.match(type_name)]
     if not matches:
         ProgrammingError(f'Unrecognized ClickHouse type {pattern} when setting formats')
     if fmt: