+++ tests/integration_tests/test_native.py
@@ -1,5 +1,9 @@
+import uuid
+from ipaddress import IPv4Address
+
 import pytest
 
+from clickhouse_connect.datatypes.format import set_default_formats, clear_default_format
 from clickhouse_connect.driver import Client
 
 
@@ -12,7 +16,7 @@ def test_low_card(test_client: Client, test_table_engine: str):
     assert len(result.result_set) == 1
 
 
-def test_json_insert(test_client: Client, test_table_engine: str):
+def test_json(test_client: Client, test_table_engine: str):
     if not test_client.min_version('22.6.1'):
         pytest.skip('JSON test skipped for old version {test_client.server_version}')
     test_client.command('DROP TABLE IF EXISTS native_json_test')
@@ -33,4 +37,25 @@ def test_json_insert(test_client: Client, test_table_engine: str):
 
 
 def test_read_formats(test_client: Client, test_table_engine: str):
-    pass
+    test_client.command('DROP TABLE IF EXISTS read_format_test')
+    test_client.command('CREATE TABLE read_format_test (key Int32, uuid UUID, fs FixedString(10), ipv4 IPv4)' +
+                        f'Engine {test_table_engine} ORDER BY key')
+    uuid1 = uuid.UUID('23E45688e89B-12D3-3273-426614174000')
+    uuid2 = uuid.UUID('77AA3278-3728-12d3-5372-000377723832')
+    row1 = (1, uuid1, '530055777k', '10.251.30.50')
+    row2 = (2, uuid2, 'short str', '10.44.75.20')
+    test_client.insert('read_format_test', [row1, row2])
+    result = test_client.query('SELECT * FROM read_format_test').result_set
+    assert result[0][1] == uuid1
+    assert result[1][3] == IPv4Address('10.44.75.20')
+    assert result[0][2] == b'\x35\x33\x30\x30\x35\x35\x37\x37\x37\x6b'
+    set_default_formats('uuid', 'string', 'ip*', 'string', 'FixedString', 'string')
+    result = test_client.query('SELECT * FROM read_format_test').result_set
+    assert result[0][1] == '23e45688-e89b-12d3-3273-426614174000'
+    assert result[1][3] == '10.44.75.20'
+    assert result[0][2] == '530055777k'
+    clear_default_format('ipv4')
+    result = test_client.query('SELECT * FROM read_format_test').result_set
+    assert result[0][1] == '23e45688-e89b-12d3-3273-426614174000'
+    assert result[1][3] == IPv4Address('10.44.75.20')
+    assert result[0][2] == '530055777k'